pyproject.toml
[project]
name = "sistemas-de-buscas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.3",
]
